Fix: moved /reservations to query controller
